// apps/genomicsdb/components/records/types.ts
export interface BaseRecord {
    id: string;
    type: "gene" | "variant" | "span" | "track";
}

// Gene record structure
export interface GeneRecord extends BaseRecord {
    type: "gene";
    symbol: string;
    name: string | null;
    synonyms: string[] | null;
    location: {
        start: number;
        end: number;
        chr: string;
        strand: "+" | "-";
    };
    hgnc_annotation: {
        agr?: string;
        ena?: string;
        lsdb?: string;
        name?: string;
        gencc?: string;
        cosmic?: string;
        mgd_id?: string;
        rgd_id?: string;
        status?: string;
        symbol?: string;
        ccds_id?: string;
        hgnc_id?: string;
        omim_id?: string;
        ucsc_id?: string;
        vega_id?: string;
        location?: string; // cytogenic location
        orphanet?: string;
        entrez_id?: string;
        prev_name?: string;
        pubmed_id?: string;
        alias_name?: string;
        gene_group?: string;
        locus_type?: string;
        locus_group?: string;
        mane_select?: string;
        prev_symbol?: string;
        uniprot_ids?: string;
        alias_symbol?: string;
        date_modified?: string;
        gene_group_id?: string;
        ensembl_gene_id?: string;
        refseq_accession?: string;
        date_name_changed?: string;
        location_sortable?: string;
        date_approved_reserved?: string;
    } | null;
    go_annotation: Array<{
        go_term_id: string;
        go_term: string;
        ontology: "BP" | "MF" | "CC"; // Biological Process, Molecular Function, Cellular Component
        evidence: Array<{
            citation: string;
            qualifier: string;
            evidence_code: string;
            go_evidence_code: string;
            annotation_source: string;
            evidence_code_qualifier: string | null;
        }>;
    }> | null;
    pathway_membership: Array<{
        pathway: string;
        pathway_id: string;
        pathway_source: string;
    }> | null;
}